+++ Release_Notes.md
@@ -123,7 +123,9 @@
 - DLogging Added a constructor that accepts connectionString so we don't have to set the properties in DLogging.
     - BLogger changed the constructor to pass the connectionString to the DLogging constructor
     - BLogger now accounts for CentralManagement
-
+- DMessages Added a constructor that accepts connectionString so we don't have to set the properties in DMessages.
+    - BMessages changed the constructor to pass the connectionString to the DMessages constructor
+    - BMessages now accounts for CentralManagement
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure